backend/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-pinecone>=0.1.0

# Document Processing
unstructured>=0.10.30
unstructured-inference>=0.7.21
pdf2image>=1.16.3
pytesseract>=0.3.10
tabulate>=0.9.0
beautifulsoup4>=4.12.2
pillow>=10.1.0
pymupdf>=1.23.7
python-pptx>=0.6.21
openpyxl>=3.1.2
pandas>=2.1.3

# Advanced Retrieval
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
cohere>=4.37  # For re-ranking

# Web Framework
Flask[async]
flask-cors
flask-limiter
gunicorn

# Async Support
aiohttp
asyncio

# External Services
firebase-admin
pinecone
twilio
google-cloud-storage
openai
anthropic
twilio
nltk
sentence-transformers

# File Processing
PyPDF2
docx2txt
python-docx
chardet
tiktoken

# Authentication
PyJWT
werkzeug

# Utilities
typing-extensions
python-dotenv
requests
